requirements.txt
pandas==1.1.3
streamlit==0.80.0
numpy==1.19.2
Pillow==8.2.0
